support software rendering on ubuntu with sdl2

+++ include/system/core/include/cutils/native_handle.h
@@ -26,7 +26,7 @@ typedef struct native_handle
     int version;        /* sizeof(native_handle_t) */
     int numFds;         /* number of file-descriptors at &data[0] */
     int numInts;        /* number of ints at &data[numFds] */
-    int data[0];        /* numFds + numInts ints */
+    uintptr_t data[0];        /* numFds + numInts ints */
 } native_handle_t;
 
 /*
